Fixed a bug where we were not soft failing on an unknown parent node for a new remote update
When we get remote updates they can be in any order and there is no guarantee that the parent node of the new node or updated node is known at this time. In this case we need to soft fail the update and just continue. Since we record all remote updates in the moveop log, we will naturally retry this update as soon as earlier updates comes in and we replay everything.